+++ app/controllers/TripController.php
@@ -1,49 +1,51 @@
 <?php
 
+use Scubawhere\Entities\Tag;
+use Scubawhere\Services\TripService;
 use Illuminate\Support\Facades\Response;
-use ScubaWhere\Services\TripService;
-use ScubaWhere\Exceptions\NotFoundException;
-use ScubaWhere\Exceptions\InvalidInputException;
+use Scubawhere\Exceptions\Http\HttpUnprocessableEntity;
 
 class TripController extends Controller {
 
-    /**
-     * Service to manage trips
-     * \ScubaWhere\Services\TripService
-     */
+    /** @var \Scubawhere\Services\TripService */
     protected $trip_service;
 
     /**
      * Response Object to create http responses
+     *
      * @var \Illuminate\Support\Facades\Response
      */
     protected $response;
 
-    /**
-     * @param TripService Injected using laravel's IOC container
-     */
     public function __construct(TripService $trip_service, Response $response) {
         $this->trip_service = $trip_service;
         $this->response = $response;
     }
 
     /**
-     * /api/trip
      * Get a single trip by ID
-     * @throws \ScubaWhere\Exceptions\NotFoundException
-     * @return json Trip model
+     *
+     * @api /api/trip
+     *
+     * @throws \Scubawhere\Exceptions\Http\HttpUnprocessableEntity
+     *
+     * @return \Scubawhere\Entities\Trip
      */
     public function getIndex() 
     {
         $id = Input::get('id');
-        if(!$id) throw new InvalidInputException(['Please provide an ID.']);
+        if(!$id) {
+            throw new HttpUnprocessableEntity(__CLASS__.__METHOD__, ['Please provide an ID.']);
+        }
         return $this->trip_service->get($id);
     }
 
     /**
-     * /api/trip/all
      * Get all trips belonging to a company
-     * @return array Collection Trip models
+     *
+     * @api /api/trip/all
+     *
+     * @return \Scubawhere\Entities\Trip
      */
     public function getAll()
     {
@@ -51,9 +53,12 @@ class TripController extends Controller {
     }
 
     /**
-     * /api/trip/all-with-trashed
      * Get all trips belonging to a company including soft deleted models
-     * @return array Collection Trip models
+     *
+     * @api /api/trip/all-with-trashed
+     *
+     * @return \Illuminate\Database\Eloquent\Collection
+     *
      */
     public function getAllWithTrashed()
     {
@@ -66,10 +71,13 @@ class TripController extends Controller {
     }
 
     /**
-     * /api/trip/add
      * Create a new trip
-     * @throws \ScubaWhere\Exceptions\InvalidInputException
-     * @return \Illuminate\Http\Response 201 Created with newly created trip
+     *
+     * @api /api/trip/add
+     *
+     * @throws \Scubawhere\Exceptions\Http\HttpNotAcceptable
+     *
+     * @return \Illuminate\Http\JsonResponse
      */
     public function postAdd()
     {
@@ -88,10 +96,14 @@ class TripController extends Controller {
     }
 
     /**
-     * /api/trip/edit
      * Edit an existing trip
-     * @throws \ScubaWhere\Exceptions\InvalidInputException
-     * @return \Illuminate\Http\Response 200 Success with updated trip
+     *
+     * @api /api/trip/edit
+     *
+     * @throws \Scubawhere\Exceptions\Http\HttpNotAcceptable
+     * @throws \Scubawhere\Exceptions\Http\HttpNotFound
+     *
+     * @return \Illuminate\Http\JsonResponse
      */
     public function postEdit()
     {
@@ -110,16 +122,22 @@ class TripController extends Controller {
     }
 
     /**
-     * /api/trip/delete
      * Delete an trip and remove it from any quotes or packages
-     * @throws \ScubaWhere\Exceptions\NotFoundException
-     * @throws Exception
+     *
+     * @api /api/trip/delete
+     *
+     * @throws \Scubawhere\Exceptions\Http\HttpNotFound
+     * @throws \Scubawhere\Exceptions\Http\HttpUnprocessableEntity
+     * @throws \Exception
+     *
      * @return \Illuminate\Http\Response 200 Success
      */
     public function postDelete()
     {
         $id = Input::get('id');
-        //if(!$id) throw new InvalidInputException(['Please provide an ID.']);
+        if(!$id) {
+            throw new HttpUnprocessableEntity(__CLASS__.__METHOD__, ['Please provide an ID.']);
+        }
         
         $this->trip_service->delete($id);
         return $this->response->json(array('status' => 'OK. Trip deleted'), 200); // 200 Success
